Webapp/requirements.txt
Flask
uWSGI
flask-mysql
mysql-connector
mysql-client
cryptography
requests
mysql-connector-python